Accepts unified input
