+++ backend/src/constants/llm.py
@@ -37,16 +37,23 @@ class ChatModels(str, Enum):
         GROQ_OPENAI_GPT_OSS_120B = "groq:openai/gpt-oss-120b"
         GROQ_LLAMA_3_3_70B_VERSATILE = "groq:llama-3.3-70b-versatile"
     if OLLAMA_BASE_URL:
-        OLLAMA_QWEN3 = "ollama:qwen3"
+        OLLAMA_QWEN3 = "ollama:qwen3-vl"
         
 def get_free_models():
-    return [
-        ChatModels.ANTHROPIC_CLAUDE_4_5_HAIKU.value,
-        ChatModels.OPENAI_GPT_5_NANO.value,
-        ChatModels.GOOGLE_GEMINI_2_5_FLASH_LITE.value,
-        ChatModels.GROQ_OPENAI_GPT_OSS_120B.value,
-        ChatModels.GROQ_LLAMA_3_3_70B_VERSATILE.value,
-    ]
+    models = []
+    if OPENAI_API_KEY:
+        models.append(ChatModels.OPENAI_GPT_5_NANO.value)
+    if ANTHROPIC_API_KEY:
+        models.append(ChatModels.ANTHROPIC_CLAUDE_4_5_HAIKU.value)
+    if GOOGLE_API_KEY:
+        models.append(ChatModels.GOOGLE_GEMINI_2_5_FLASH_LITE.value)
+    if GROQ_API_KEY:
+        models.append(ChatModels.GROQ_OPENAI_GPT_OSS_120B.value)
+    if GROQ_API_KEY:
+        models.append(ChatModels.GROQ_LLAMA_3_3_70B_VERSATILE.value)
+    if OLLAMA_BASE_URL:
+        models.append(ChatModels.OLLAMA_QWEN3.value)
+    return models
 
 
 def get_system_prompt():
