+++ backend/src/utils/auth.py
@@ -8,7 +8,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from src.constants.llm import get_free_models
 from src.repos.user_repo import UserRepo
 from src.constants import JWT_SECRET_KEY, JWT_ALGORITHM, JWT_TOKEN_EXPIRE_MINUTES
-from src.schemas.entities import LLMRequest, LLMStreamRequest
+from src.schemas.entities import LLMRequest
 from src.schemas.models import User
 from src.services.db import get_async_db
 from src.utils.logger import logger
@@ -45,7 +45,7 @@ def is_authorized_model(model: str) -> bool:
 
 async def get_optional_user(
     request: Request,
-    params: LLMRequest | LLMStreamRequest,
+    params: LLMRequest,
     credentials: Optional[HTTPAuthorizationCredentials] = Depends(security),
     db: AsyncSession = Depends(get_async_db),
 ) -> Optional[User]:
