+++ frontend/src/lib/services/threadService.ts
@@ -84,21 +84,28 @@ export const alterSystemPrompt = async (payload: ThreadPayload) => {
 };
 
 type MessageContent = string | Array<{ type: string; [key: string]: any }>;
-
+type Messages = { role: string; content: MessageContent; [key: string]: any }[];
+type Input = { messages: Messages };
+type Metadata = { thread_id?: string; checkpoint_id?: string; [key: string]: any };
+type A2A = { [key: string]: any };
+type MCP = { [key: string]: any };
+type Tools = string[];
+type Subagents = Agent[];
+type Presidio = {
+	analyze?: boolean;
+	anonymize?: boolean;
+	redact?: boolean;
+};
 interface StreamThreadPayload {
 	system?: string;
-	messages: { role: string; content: MessageContent; [key: string]: any }[];
+	input: Input;
 	model: string;
-	metadata: { thread_id?: string; checkpoint_id?: string; [key: string]: any };
-	a2a?: object;
-	mcp?: object;
-	tools?: string[];
-	subagents?: Agent[];
-	presidio?: {
-		analyze?: boolean;
-		anonymize?: boolean;
-		redact?: boolean;
-	};
+	metadata: Metadata;
+	a2a?: A2A;
+	mcp?: MCP;
+	tools?: Tools;
+	subagents?: Subagents;
+	presidio?: Presidio;
 }
 
 export const streamThread = (payload: StreamThreadPayload): SSE => {
